fix(external-models): group manual model list by route
