+++ web/src/locales/en.json
@@ -2293,7 +2293,13 @@
     "placeholder": "gpt-5\nclaude-sonnet-4\ngemini-2.5-pro",
     "saveHint": "This list is used only while the External model list switch is enabled in Settings.",
     "modelCount": "{{count}} models",
-    "empty": "No external models configured. With the switch enabled, model-list APIs will return an empty list."
+    "empty": "No external models configured. With the switch enabled, model-list APIs will return an empty list.",
+    "groupedTitle": "Grouped by route",
+    "groupedDesc": "Groups the manually maintained external model names above by enabled route, provider, and model mapping.",
+    "groupCount": "{{count}} groups",
+    "globalScope": "Global",
+    "uncategorizedTitle": "No matching route",
+    "uncategorizedDesc": "These models are still saved in the external model list, but do not currently match an enabled route."
   },
   "proxies": {
     "title": "Proxies",
