+++ web/src/locales/zh.json
@@ -2289,7 +2289,13 @@
     "placeholder": "gpt-5\nclaude-sonnet-4\ngemini-2.5-pro",
     "saveHint": "仅当设置里的「外部模型列表」开关开启时，此列表才会生效。",
     "modelCount": "{{count}} 个模型",
-    "empty": "还没有配置外部模型。开关开启时，模型列表接口会返回空列表。"
+    "empty": "还没有配置外部模型。开关开启时，模型列表接口会返回空列表。",
+    "groupedTitle": "按路由分类预览",
+    "groupedDesc": "根据已启用路由、Provider 和模型映射，把上面手动维护的外部模型名分组显示。",
+    "groupCount": "{{count}} 个分类",
+    "globalScope": "全局",
+    "uncategorizedTitle": "未匹配到路由",
+    "uncategorizedDesc": "这些模型仍会保存在外部模型列表中，但当前未匹配到已启用路由。"
   },
   "proxies": {
     "title": "代理",
